refactor(Hoster): moved validateExcel to sep class
GH-31: created new validator package and moved validate excel logic to new class ExcelFileValidator
