+++ hoster/src/main/java/com/excel/hoster/controller/UploadExcelControllerApi.java
@@ -2,10 +2,11 @@ package com.excel.hoster.controller;
 
 
 import com.excel.hoster.repository.entity.ExcelFileEntity;
-import com.excel.dto.ExcelFileDTO;
+import com.excel.hoster.dto.ExcelFileDTO;
 import com.excel.hoster.service.ExcelFileService;
 import com.excel.hoster.repository.ExcelRepository;
 import com.excel.hoster.service.ObjectResponse;
+import com.excel.hoster.validator.ExcelFileValidator;
 import jakarta.validation.Valid;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -40,7 +41,7 @@ public class UploadExcelControllerApi {
             BindingResult bindingResult, Model model)
             throws IOException {
 
-        ExcelFileService.validateExcel(bindingResult, file);
+        ExcelFileValidator.validateExcel(bindingResult, file);
 
         if (bindingResult.hasErrors()) {
             return ResponseEntity.badRequest().body(bindingResult.getAllErrors());
