+++ hoster/src/main/java/com/excel/hoster/controller/UploadExcelControllerWeb.java
@@ -1,9 +1,10 @@
 package com.excel.hoster.controller;
 
 import com.excel.hoster.repository.entity.ExcelFileEntity;
-import com.excel.dto.ExcelFileDTO;
+import com.excel.hoster.dto.ExcelFileDTO;
 import com.excel.hoster.service.ExcelFileService;
 import com.excel.hoster.repository.ExcelRepository;
+import com.excel.hoster.validator.ExcelFileValidator;
 import jakarta.validation.Valid;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -48,7 +49,7 @@ public class UploadExcelControllerWeb {
             @RequestParam(name="file",required = false) MultipartFile file,
             BindingResult bindingResult, Model model) throws IOException {
 
-        ExcelFileService.validateExcel(bindingResult,file);
+        ExcelFileValidator.validateExcel(bindingResult,file);
 
         ExcelFileEntity excelFile = new ExcelFileEntity(excelFileDTO.getDefinitionName(), excelFileDTO.getBrickName(), excelFileDTO.getAttributeName(),file.getOriginalFilename(), file.getBytes());
         model.addAttribute("excelFile", excelFile);
