+++ hoster/src/main/java/com/excel/hoster/dto/ExcelFileDTO.java
@@ -1,4 +1,4 @@
-package com.excel.dto;
+package com.excel.hoster.dto;
 
 import jakarta.validation.constraints.NotBlank;
