+++ hoster/src/test/java/com/excel/hoster/service/ExcelFileServiceTest.java
@@ -1,6 +1,7 @@
 package com.excel.hoster.service;
 
 import com.excel.hoster.service.ExcelFileService;
+import com.excel.hoster.validator.ExcelFileValidator;
 import org.junit.jupiter.api.*;
 
 import org.mockito.Mock;
@@ -43,7 +44,7 @@ public class ExcelFileServiceTest {
         @DisplayName("it should not give an error if the file is valid")
         @Test
         void testExcelFileValid() {
-            ExcelFileService.validateExcel(bindingResult,mockFile);
+            ExcelFileValidator.validateExcel(bindingResult,mockFile);
             verify(bindingResult,never()).rejectValue(anyString(),anyString(),anyString());
         }
 
@@ -53,7 +54,7 @@ public class ExcelFileServiceTest {
         void testExcelFileTooBig() {
             when(mockFile.getSize()).thenReturn(21L*1024*1024);
 
-            ExcelFileService.validateExcel(bindingResult,mockFile);
+            ExcelFileValidator.validateExcel(bindingResult,mockFile);
             verify(bindingResult,times(1)).rejectValue("excelFile","400","Excel file must be less than 20MB");
         }
 
@@ -63,7 +64,7 @@ public class ExcelFileServiceTest {
         void testExcelFileFileNotExcel() {
             when(mockFile.getOriginalFilename()).thenReturn("test.ppt");
 
-            ExcelFileService.validateExcel(bindingResult,mockFile);
+            ExcelFileValidator.validateExcel(bindingResult,mockFile);
             verify(bindingResult,times(1)).rejectValue("excelFile","400","File must be .xls or .xlsx");
         }
 
@@ -72,7 +73,7 @@ public class ExcelFileServiceTest {
         @Test
         void testExcelFileIsNull() {
 
-            ExcelFileService.validateExcel(bindingResult,null);
+            ExcelFileValidator.validateExcel(bindingResult,null);
             verify(bindingResult,times(1)).rejectValue("excelFile","400","excelFile is mandatory");
